Reap connections based on owning-thread death
.. not a general timeout.

Now, if a thread checks out a connection then dies, we can immediately
recover that connection and re-use it.

This should alleviate the pool exhaustion discussed in #12867. More
importantly, it entirely avoids the potential issues of the reaper
attempting to check whether connections are still active: as long as the
owning thread is alive, the connection is its business alone.

As a no-op reap is now trivial (only entails checking a thread status
per connection), we can also perform one in-line any time we decide to
sleep for a connection.
(cherry picked from commit 9e457a8654fa89fe329719f88ae3679aefb21e56)

Conflicts:
	activerecord/CHANGELOG.md
+ (babbtx) didn't take this change from 9e457a8654fa89fe329719f88ae3679aefb21e56)